requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.4.0
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
datasets==4.1.1
dill==0.4.0
filelock==3.13.1
frozenlist==1.8.0
fsspec==2024.6.1
gensim==4.3.3
hdbscan==0.8.40
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.4
joblib==1.5.2
llvmlite==0.45.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.3
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==25.0
pandas==2.3.3
pillow==11.0.0
propcache==0.4.0
psutil==7.1.0
pyarrow==21.0.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.13.1
sentence-transformers==5.1.1
setuptools==70.2.0
six==1.17.0
smart_open==7.3.1
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.56.2
triton==3.3.1
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
